+++ app/(home)/user/settings/delete.tsx
@@ -1,0 +1,5 @@
+import { DeleteScreen } from 'appdeptus/modules/user/Screens'
+
+const DeleteRoute = () => <DeleteScreen />
+
+export default DeleteRoute