+++ components/ScreenContainer/ScreenContainer.tsx
@@ -7,13 +7,17 @@ import { selectThemeName } from '../store'
 import { themeColors, VStack } from '../ui'
 
 type ScreenContainerProps = {
+  hideBottomGradient?: boolean
   safeAreaInsets?: ('top' | 'bottom')[]
 } & ComponentProps<typeof VStack>
 
 const ScreenContainer = ({
   children,
   className,
+
+  hideBottomGradient,
   safeAreaInsets,
+
   ...props
 }: PropsWithChildren<ScreenContainerProps>) => {
   const { top, bottom } = useSafeAreaInsets()
@@ -34,18 +38,20 @@ const ScreenContainer = ({
       >
         {children}
       </VStack>
-      <LinearGradient
-        colors={[
-          `${themeColors[themeName].primary[950]}00`,
-          themeColors[themeName].primary[950]
-        ]}
-        style={{
-          position: 'absolute',
-          height: 40,
-          width: '100%',
-          bottom: safeAreaInsets?.includes('bottom') ? bottom - 1 : 0 // this shitty -1 is a workaround to make sure that the gradient meats the safe area correctly
-        }}
-      />
+      {!hideBottomGradient ? (
+        <LinearGradient
+          colors={[
+            `${themeColors[themeName].primary[950]}00`,
+            themeColors[themeName].primary[950]
+          ]}
+          style={{
+            position: 'absolute',
+            height: 40,
+            width: '100%',
+            bottom: safeAreaInsets?.includes('bottom') ? bottom - 1 : 0 // this shitty -1 is a workaround to make sure that the gradient meats the safe area correctly
+          }}
+        />
+      ) : null}
     </>
   )
 }