+++ modules/user/Screens/index.ts
@@ -1,2 +1,4 @@
+export { default as DeleteScreen } from './Delete'
 export { default as EditNameScreen } from './EditName'
+export { default as SettingsScreen } from './Settings'
 export { default as UserScreen } from './User'